// gui/frontend/src/store/redux-query/models/Queue.ts
// tslint:disable
/**
 * Onyks Wagon Location API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import { Project, ProjectFromJSON, ProjectToJSON } from ".";

/**
 *
 * @export
 * @interface Queue
 */
export interface Queue {
  /**
   *
   * @type {string}
   * @memberof Queue
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof Queue
   */
  name: string;
  /**
   *
   * @type {Project}
   * @memberof Queue
   */
  project?: Project;
}

export function QueueFromJSON(json: any): Queue {
  return {
    id: json["id"],
    name: json["name"],
    project: !exists(json, "project")
      ? undefined
      : ProjectFromJSON(json["project"]),
  };
}

export function QueueToJSON(value?: Queue): any {
  if (value === undefined) {
    return undefined;
  }
  return {
    id: value.id,
    name: value.name,
    project: ProjectToJSON(value.project),
  };
}